src/gymnasium_env/__init__.py
"""
Gymnasium Environment Wrapper for Minecraft
Replaces MineDojo with gymnasium-compatible interface
"""

import gymnasium as gym
from gymnasium import spaces
import numpy as np
from typing import Dict, Any, Tuple, Optional
import cv2


class MinecraftGymnasiumEnv(gym.Env):
    """
    Gymnasium wrapper for Minecraft environment
    Removes Java dependencies and provides clean interface
    """
    
    def __init__(self, 
                 name: str = "Plains",
                 img_size: Tuple[int, int] = (640, 480),
                 rgb_only: bool = False,
                 **kwargs):
        super().__init__()
        
        self.name = name
        self.img_size = img_size
        self.rgb_only = rgb_only
        
        # Define action space - simplified from MineDojo
        self.action_space = self._create_action_space()
        
        # Define observation space
        self.observation_space = self._create_observation_space()
        
        # Internal state
        self._current_step = 0
        self._max_steps = kwargs.get('max_steps', 1000)
        self._inventory = {}
        
    def _create_action_space(self):
        """Create simplified action space without Java dependencies"""
        # Simplified action space for testing
        # In real implementation, this would interface with Minecraft MDK
        return spaces.Dict({
            'movement': spaces.Discrete(4),  # forward, back, left, right
            'camera': spaces.Box(low=-1.0, high=1.0, shape=(2,), dtype=np.float32),
            'use': spaces.Discrete(2),  # use item
            'attack': spaces.Discrete(2),  # attack
            'jump': spaces.Discrete(2),  # jump
        })
    
    def _create_observation_space(self):
        """Create observation space"""
        obs_space = {
            'rgb': spaces.Box(
                low=0, high=255, 
                shape=(*self.img_size, 3), 
                dtype=np.uint8
            ),
            'inventory': spaces.Dict({
                # Simplified inventory space
                'wood': spaces.Discrete(64),
                'stone': spaces.Discrete(64),
                'iron': spaces.Discrete(64),
            })
        }
        
        if not self.rgb_only:
            obs_space['depth'] = spaces.Box(
                low=0, high=255,
                shape=self.img_size,
                dtype=np.uint8
            )
        
        return spaces.Dict(obs_space)
    
    def reset(self, seed: Optional[int] = None, options: Optional[Dict] = None):
        """Reset environment"""
        super().reset(seed=seed)
        
        self._current_step = 0
        self._inventory = {'wood': 0, 'stone': 0, 'iron': 0}
        
        # Generate dummy observation for testing
        obs = self._get_observation()
        info = {'step': self._current_step}
        
        return obs, info
    
    def step(self, action: Dict[str, Any]):
        """Execute action in environment"""
        self._current_step += 1
        
        # Process action (dummy implementation)
        self._process_action(action)
        
        # Get new observation
        obs = self._get_observation()
        
        # Calculate reward (dummy)
        reward = self._calculate_reward()
        
        # Check if done
        terminated = self._current_step >= self._max_steps
        truncated = False
        
        info = {
            'step': self._current_step,
            'inventory': self._inventory.copy()
        }
        
        return obs, reward, terminated, truncated, info
    
    def _process_action(self, action: Dict[str, Any]):
        """Process action (dummy implementation)"""
        # In real implementation, this would send actions to Minecraft MDK
        pass
    
    def _get_observation(self):
        """Get current observation"""
        # Generate dummy RGB image
        rgb = np.random.randint(0, 255, (*self.img_size, 3), dtype=np.uint8)
        
        obs = {
            'rgb': rgb,
            'inventory': self._inventory.copy()
        }
        
        if not self.rgb_only:
            # Generate dummy depth image
            depth = np.random.randint(0, 255, self.img_size, dtype=np.uint8)
            obs['depth'] = depth
        
        return obs
    
    def _calculate_reward(self):
        """Calculate reward (dummy implementation)"""
        return 0.0
    
    def render(self, mode='rgb_array'):
        """Render environment"""
        if mode == 'rgb_array':
            return self._get_observation()['rgb']
        else:
            raise NotImplementedError(f"Render mode {mode} not supported")
    
    def close(self):
        """Close environment"""
        pass


class MinecraftEnvRegistry:
    """Registry for different Minecraft environments"""
    
    _environments = {
        'Plains': MinecraftGymnasiumEnv,
        'Forest': MinecraftGymnasiumEnv,
        'Mountain': MinecraftGymnasiumEnv,
    }
    
    @classmethod
    def make(cls, name: str, **kwargs):
        """Create environment instance"""
        if name not in cls._environments:
            raise ValueError(f"Environment {name} not found")
        
        return cls._environments[name](name=name, **kwargs)
    
    @classmethod
    def register(cls, name: str, env_class):
        """Register new environment"""
        cls._environments[name] = env_class


# Compatibility layer for MineDojo API
def MineDojoEnv(name: str = "Plains", **kwargs):
    """
    Compatibility function to replace MineDojo environment
    """
    return MinecraftEnvRegistry.make(name, **kwargs)
